RAG_Finance/requirements.txt
   # Create a new file called requirements_minimal.txt with:
   fastapi==0.95.1
   uvicorn==0.22.0
   python-dotenv==1.0.0
   openai==0.27.8
   pinecone-client==2.2.2
   redis==4.5.5